+++ src/instruments/src/EFB/Utils/AircraftVersionChecker.ts
@@ -1,12 +1,11 @@
-/* eslint-disable no-console */
 // Copyright (c) 2022 FlyByWire Simulations
 // SPDX-License-Identifier: GPL-3.0
 
+/* eslint-disable no-console */
 import Compare from 'semver/functions/compare';
 import { CommitInfo, GitVersions, ReleaseInfo } from '@flybywiresim/api-client';
-// jest in test-js.sh requires relative path  can't handle "@shared" alias
-// noinspection ES6PreferShortImport
-import { PopUp } from '../../../../shared/src/popup';
+import { PopUp } from '@shared/popup';
+import { t } from '../translation';
 
 /**
  * Contains the a32nx_build_info.json file's information in a structured way.
@@ -128,19 +127,13 @@ export class AircraftVersionChecker {
     public static getVersionInfo(versionString: string): VersionInfoData {
         const matchBuildInfo = versionString.match(/^v?((\d+)\.(\d+)\.(\d+))-(.*)\.(.{7})$/);
         if (matchBuildInfo) {
-            const version = matchBuildInfo[1];
-            const major = parseInt(matchBuildInfo[2], 10);
-            const minor = parseInt(matchBuildInfo[3], 10);
-            const patch = parseInt(matchBuildInfo[4], 10);
-            const branch = matchBuildInfo[5];
-            const commit = matchBuildInfo[6];
             return {
-                version,
-                major,
-                minor,
-                patch,
-                branch,
-                commit,
+                version: matchBuildInfo[1],
+                major: parseInt(matchBuildInfo[2], 10),
+                minor: parseInt(matchBuildInfo[3], 10),
+                patch: parseInt(matchBuildInfo[4], 10),
+                branch: matchBuildInfo[5],
+                commit: matchBuildInfo[6],
             };
         }
         throw new Error('Invalid version format');
@@ -242,12 +235,16 @@ export class AircraftVersionChecker {
     private static showVersionPopup(branchName, currentVersion, releaseVersion) {
         const popup = new PopUp();
         popup.showInformation(
-            'NEW VERSION AVAILABLE',
+            t('VersionCheck.Title'),
             `<div style="font-size: 100%; text-align: left;">
-                     You are using ${branchName} version:<br/><strong>${currentVersion}</strong><br/><br/> 
-                     Latest ${branchName} version is:<br /><strong>${releaseVersion}</strong><br/><br/>
-                     Please update your aircraft using the FlyByWire Installer.
-                     </div>`,
+                        ${t('VersionCheck.CurrentVersionText', [{ edition: branchName }])}<br>
+                        <strong>${currentVersion}</strong><br><br>
+                                        
+                        ${t('VersionCheck.LatestVersionText', [{ edition: branchName }])}<br>
+                        <strong>${releaseVersion}</strong><br/><br/>
+                        
+                        ${t('VersionCheck.RecommendationText')}
+                    </div>`,
             'normal',
             () => {},
         );